requirements.txt
strands-agents
strands-agents[ollama]
python-dotenv
argparse
mcp-python-client
readline
coverage==7.5.0
